+++ portal-impl/src/com/liferay/portlet/StateAwareResponseImpl.java
@@ -24,6 +24,7 @@ import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.portlet.PortletQNameUtil;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
+import com.liferay.portlet.internal.EventImpl;
 
 import java.io.Serializable;
