+++ portal-impl/src/com/liferay/portlet/internal/EventImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet;
+package com.liferay.portlet.internal;
 
 import java.io.Serializable;
